+++ app/javascript/src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationComponents.js
@@ -1,10 +1,19 @@
+/*
+The variations grid itself: the row-level columns every variation has regardless of its scheme, the
+column groups and their headers, visibility, ordering and persistence, and the AG Grid instance that
+holds them.
+
+What a column of the scheme shows - materials and reaction fields - lives in
+ReactionVariationSchemaComponents.
+*/
 import React, {
-  createContext, useCallback, useContext, useEffect, useMemo, useRef, useState
+  useCallback, useContext, useEffect, useMemo, useRef, useState
 } from 'react';
 import PropTypes from 'prop-types';
+import { Select } from 'src/components/common/Select';
 import { AgGridReact } from 'ag-grid-react';
 import {
-  Button, ButtonGroup, Form, InputGroup, OverlayTrigger, Popover
+  Button, ButtonGroup, Form, OverlayTrigger, Popover
 } from 'react-bootstrap';
 import ReorderableList from 'src/components/common/ReorderableList';
 import Reaction from 'src/models/Reaction';
@@ -12,452 +21,31 @@ import { permitOn } from 'src/components/common/uis';
 import DragHandle from 'src/components/common/DragHandle';
 import DeleteButton from 'src/components/common/DeleteButton';
 import AppModal from 'src/components/common/AppModal';
-import { isSbmmSample } from 'src/utilities/ElementUtils';
-import { metPrefSymbols } from 'src/utilities/metricPrefix';
-import {
-  getMetricMolConc, metricPrefixesMol, metricPrefixesMolConc
-} from 'src/utilities/MetricsUtils';
 import { AnalysesCell } from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsAnalyses';
 import ReactionUpdateHandler from 'src/apps/mydb/elements/details/reactions/schemeTab/ReactionUpdateUtils';
-import MaterialHandler from 'src/apps/mydb/elements/details/reactions/schemeTab/material/MaterialUtils';
-import { MATERIAL_HEADER } from 'src/apps/mydb/elements/details/reactions/schemeTab/MaterialGroup';
 import {
   getInitialColumnState,
-  persistColumnState
+  persistColumnState,
+  GROUP_ID_SEPARATOR, getReactionSegments
 } from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsUtils';
-import REACTION_FIELDS from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationReactionFields';
+import VariationsGridContext
+  from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsGridContext';
 import {
-  CoefficientField,
-  DrySolventCheckBox,
-  EquivalentOrYield,
-  GaseousInputFields,
-  MassField,
-  MaterialActivity,
-  MaterialAmountMol,
-  MaterialConcentration,
-  MaterialLoading,
-  MaterialNameWithIupac,
-  MaterialRef,
-  MaterialVolume,
-  SwitchTargetReal,
-  VolumeRatio,
-  VOLUME_METRIC_PREFIXES,
-  volumeMetricPrefix,
-  molMetricPrefix
-} from 'src/apps/mydb/elements/details/reactions/schemeTab/material/MaterialComponents';
+  schemaBuildColumnGroups
+} from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationSchemaComponents';
+import {
+  segmentBuildColumnGroups
+} from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationSegmentComponents';
 
-const MAT_GROUPS = ['starting_materials', 'reactants', 'solvents', 'products'];
-const MAT_GROUP_TITLES = {
-  starting_materials: 'Starting material',
-  reactants: 'Reactant',
-  solvents: 'Solvent',
-  products: 'Product',
-};
-const MASS_METRIC_PREFIXES = ['m', 'n', 'u'];
 // Material group names contain underscores, so the group id needs a separator that cannot collide.
-const GROUP_ID_SEPARATOR = '::';
+
 // Pseudo material groups, so the row-level and reaction-level columns are toggled like the rest.
-const REACTION_FIELDS_GROUP = 'reaction_fields';
+
 const VARIATION_GROUP = 'variation_fields';
 const ANALYSES_GROUP = 'analyses_fields';
 // Marks the material name cells that follow the horizontal scroll inside their own group.
 const STICKY_NAME_CLASS = 'variations-sticky-name';
 const STICKY_NAME_FLOATING_CLASS = 'variations-sticky-name--floating';
-
-/*
-Carries the per-row update handlers into the cells.
-
-The context value is deliberately NOT memoized: AG Grid does not re-render cell components when the
-surrounding React tree re-renders, but context updates do reach them (portals keep the React tree
-intact). Recreating the value on every render is what keeps every cell in sync with the mutated
-Reaction models after an edit, without going through `api.refreshCells`, which would tear down and
-rebuild the cell component and steal focus from the input being typed into.
-*/
-const VariationsGridContext = createContext({
-  variations: [],
-  columnUnits: {},
-  setColumnUnit: () => {},
-  getRowHandler: () => null,
-  setActiveVariation: () => {},
-  onGroupChange: () => {},
-});
-
-const isEmptyValue = (v) => v === null || v === undefined || Number.isNaN(v) || v === 0;
-
-// Same condition the scheme tab uses to decide whether a product gets its gas phase row.
-const isGasProductMaterial = (reaction, material) => (
-  !!reaction?.gaseous && material?.gas_type === 'gas'
-);
-
-// Mirrors the mass metric prefix resolution of the scheme tab's GeneralMaterial row.
-const massMetricPrefix = (material, isSbmm) => {
-  if (isSbmm) {
-    return material.reactionSchemeMetricPrefix(material.amount_as_used_mass_unit);
-  }
-  if (
-    material.metrics
-    && material.metrics.length > 2
-    && MASS_METRIC_PREFIXES.indexOf(material.metrics[0]) > -1
-  ) {
-    return material.metrics[0];
-  }
-  return 'm';
-};
-
-const SolventLabel = ({ mh }) => {
-  const { material, reaction, materialGroup } = mh;
-
-  return (
-    <InputGroup>
-      <Form.Control
-        disabled={!permitOn(reaction)}
-        type="text"
-        size="sm"
-        value={material.external_label ?? ''}
-        placeholder={
-          mh.isSbmm
-            ? (material.name || material.short_label || '')
-            : (material.molecule?.iupac_name || '')
-        }
-        onChange={(event) => mh.handler.externalLabelChange(event)}
-      />
-      <Button
-        disabled={materialGroup === 'purification_solvents' || !permitOn(reaction)}
-        onClick={() => mh.handler.externalLabelCompleted()}
-        size="sm"
-      >
-        <i className="fa fa-refresh" />
-      </Button>
-    </InputGroup>
-  );
-};
-
-SolventLabel.propTypes = {
-  mh: PropTypes.instanceOf(MaterialHandler).isRequired,
-};
-
-const PlainValue = ({ children }) => <span className="px-1">{children}</span>;
-
-PlainValue.propTypes = {
-  children: PropTypes.node,
-};
-
-const NAME_FIELD = {
-  key: 'name',
-  header: 'Material',
-  width: 230,
-  // Marks the cell for the sticky handling in updateStickyNames.
-  sticky: true,
-  render: (mh, { index }) => (
-    <MaterialNameWithIupac mh={mh} index={index} withStickyName={false} />
-  ),
-};
-
-const GENERAL_MATERIAL_SETTIGS_FIELDS = [
-  NAME_FIELD,
-  {
-    key: 'ref',
-    header: MATERIAL_HEADER.ref,
-    width: 60,
-    render: (mh) => <MaterialRef mh={mh} />,
-  },
-  {
-    key: 'tr',
-    header: MATERIAL_HEADER.tr,
-    width: 64,
-    render: (mh) => <SwitchTargetReal mh={mh} />,
-  },
-  {
-    key: 'coefficient',
-    header: MATERIAL_HEADER.reaction_coefficient,
-    width: 90,
-    render: (mh) => (mh.isSbmm ? null : <CoefficientField mh={mh} />),
-  }
-];
-
-/*
-One entry per grid column. `render` is handed a MaterialHandler bound to the material of that row,
-so every column reuses the very same input component the scheme tab renders.
-*/
-const GENERAL_MATERIAL_AMOUNT_FIELDS = [
-  {
-    key: 'mass',
-    header: MATERIAL_HEADER.mass,
-    width: 150,
-    unitToggle: { unit: 'g', prefixes: MASS_METRIC_PREFIXES, prefixOf: massMetricPrefix },
-    render: (mh) => (
-      <MassField
-        mh={mh}
-        metric={massMetricPrefix(mh.material, mh.isSbmm)}
-        metricPrefixes={MASS_METRIC_PREFIXES}
-      />
-    ),
-  },
-  {
-    key: 'volume',
-    header: MATERIAL_HEADER.vol,
-    width: 150,
-    unitToggle: { unit: 'l', prefixes: VOLUME_METRIC_PREFIXES, prefixOf: volumeMetricPrefix },
-    render: (mh) => <MaterialVolume mh={mh} className="reaction-material__volume-data" />,
-  },
-  {
-    key: 'amount',
-    header: MATERIAL_HEADER.amount,
-    width: 150,
-    unitToggle: { unit: 'mol', prefixes: metricPrefixesMol, prefixOf: molMetricPrefix },
-    render: (mh) => <MaterialAmountMol mh={mh} />,
-  },
-  {
-    key: 'molar_mass',
-    header: MATERIAL_HEADER.molar_mass,
-    width: 120,
-    render: (mh) => (mh.isSbmm
-      ? <MaterialActivity mh={mh} />
-      : <PlainValue>{mh.molarWeightValue(true)}</PlainValue>),
-  },
-  {
-    key: 'density',
-    header: MATERIAL_HEADER.density,
-    width: 80,
-    render: (mh) => <PlainValue>{mh.material.has_density ? mh.material.density : 'undefined'}</PlainValue>,
-  },
-  {
-    key: 'purity',
-    header: MATERIAL_HEADER.purity,
-    width: 80,
-    render: (mh) => {
-      const { purity } = mh.material;
-      return <PlainValue>{(purity === null || purity === undefined || purity === '') ? 0 : purity}</PlainValue>;
-    },
-  },
-  {
-    key: 'loading',
-    header: MATERIAL_HEADER.loading,
-    width: 130,
-    // Only rendered for reactions with polymers, matching the scheme tab's loading column.
-    requiresLoadingColumn: true,
-    render: (mh, { showLoadingColumn }) => (
-      <MaterialLoading mh={mh} showLoadingColumn={showLoadingColumn} />
-    ),
-  },
-  {
-    key: 'concn',
-    header: MATERIAL_HEADER.concn,
-    width: 150,
-    unitToggle: {
-      unit: 'mol/l',
-      prefixes: metricPrefixesMolConc,
-      prefixOf: (material) => getMetricMolConc(material),
-    },
-    render: (mh) => <MaterialConcentration mh={mh} />,
-  },
-  {
-    key: 'eq',
-    header: MATERIAL_HEADER.eq,
-    width: 150,
-    render: (mh, { displayYieldField }) => (
-      <EquivalentOrYield mh={mh} displayYieldField={displayYieldField} />
-    ),
-  },
-];
-
-const SOLVENT_FIELDS = [
-  NAME_FIELD,
-  {
-    key: 'dry_solvent',
-    header: 'Dry',
-    width: 60,
-    render: (mh) => <DrySolventCheckBox mh={mh} />,
-  },
-  {
-    key: 'tr',
-    header: MATERIAL_HEADER.tr,
-    width: 64,
-    render: (mh) => <SwitchTargetReal mh={mh} />,
-  },
-  {
-    key: 'label',
-    header: 'Label',
-    width: 220,
-    render: (mh) => <SolventLabel mh={mh} />,
-  },
-  {
-    key: 'volume',
-    header: MATERIAL_HEADER.vol,
-    width: 150,
-    render: (mh) => <MaterialVolume mh={mh} className="reaction-material__solvent-volume-data" />,
-  },
-  {
-    key: 'ratio',
-    header: 'Ratio',
-    width: 90,
-    render: (mh) => <VolumeRatio mh={mh} />,
-  },
-];
-
-/*
-Gas phase inputs, which the scheme tab shows as an extra row under a gaseous product. Only the three
-editable ones are columns here; turnover number and turnover frequency are derived and read-only in
-GaseousInputFields, so they would just be dead columns.
-
-The columns exist only for product slots where some variation actually has a gaseous product - see
-FIELDS_BY_GROUP.products below - and within them a row renders nothing unless its own product is the
-gas one (`isGasProduct`), since a variation may well have turned the gas mode off.
-*/
-const GAS_PHASE_FIELDS = [
-  // ppm has no unit to switch, so it gets no header unit button either.
-  { key: 'gas_time', header: 'Time', gasField: 'time', unitSwitchable: true },
-  { key: 'gas_temperature', header: 'Temp', gasField: 'temperature', unitSwitchable: true },
-  { key: 'gas_ppm', header: 'ppm', gasField: 'part_per_million' },
-].map(({
-  key, header, gasField, unitSwitchable
-}) => ({
-  key,
-  header,
-  width: 150,
-  ...(unitSwitchable ? { unitToggle: { gasField } } : {}),
-  render: (mh, { isGasProduct }) => (
-    isGasProduct ? <GaseousInputFields mh={mh} field={gasField} /> : null
-  ),
-}));
-
-const GENERAL_MATERIAL_FIELDS = [...GENERAL_MATERIAL_SETTIGS_FIELDS, ...GENERAL_MATERIAL_AMOUNT_FIELDS];
-const GENERAL_GAS_MATERIAL_FIELDS = [
-  ...GENERAL_MATERIAL_SETTIGS_FIELDS,
-  ...GAS_PHASE_FIELDS,
-  ...GENERAL_MATERIAL_AMOUNT_FIELDS
-];
-
-const FIELDS_BY_GROUP = {
-  starting_materials: () => GENERAL_MATERIAL_FIELDS,
-  reactants: () => GENERAL_MATERIAL_FIELDS,
-  products: (gasType) => gasType ? GENERAL_GAS_MATERIAL_FIELDS : GENERAL_MATERIAL_FIELDS,
-  solvents: () => SOLVENT_FIELDS,
-};
-
-/*
-Builds the MaterialHandler for one material of one row. `matGroup` may be null (no slot in view),
-in which case there is nothing to render.
-*/
-const useMaterialHandler = (data, matGroup, sampleIdx) => {
-  const { getRowHandler } = useContext(VariationsGridContext);
-  const variationReaction = data?.data ?? null;
-  const material = (matGroup && variationReaction?.[matGroup]?.[sampleIdx]) || null;
-
-  // Only the equivalent/weight-percentage selector reads this, and it is per material, so cell-local
-  // state is the right scope for it.
-  const [fieldToShow, setFieldToShow] = useState(
-    () => (material && !isEmptyValue(material.weight_percentage) ? 'weight percentage' : 'molar mass')
-  );
-
-  const rowHandler = variationReaction ? getRowHandler(data) : null;
-  // Read as a dependency rather than only inside the factory, so a change of the equivalent lock
-  // rebuilds the handler instead of leaving the disabled states stale.
-  const lockEquivColumn = rowHandler ? rowHandler.lockEquivColumn : false;
-
-  return useMemo(() => {
-    if (!material || !rowHandler) {
-      return null;
-    }
-    return new MaterialHandler({
-      index: sampleIdx,
-      material,
-      reaction: variationReaction,
-      materialGroup: matGroup,
-      onChange: rowHandler.handleMaterialsChange,
-      setFieldToShow,
-      fieldToShow,
-      mixtureComponents: [],
-      setMixtureComponents: () => {},
-      lockEquivColumn,
-    });
-  }, [material, variationReaction, matGroup, rowHandler, fieldToShow, lockEquivColumn]);
-};
-
-/*
-Renders a single input of a single material. `matGroup`/`sampleIdx`/`field` are fixed per column,
-the row supplies the variation reaction.
-*/
-const MaterialFieldCell = ({
-  data, colId, matGroup, sampleIdx, field
-}) => {
-  const { columnUnits } = useContext(VariationsGridContext);
-  const mh = useMaterialHandler(data, matGroup, sampleIdx);
-
-  if (!mh) {
-    return null;
-  }
-
-  const variationReaction = mh.reaction;
-  const input = field.render(mh, {
-    index: sampleIdx + 1,
-    showLoadingColumn: !!variationReaction.hasPolymers(),
-    displayYieldField: variationReaction.products.every(
-      (product) => !(product.conversion_rate && product.conversion_rate !== 0)
-    ),
-    isGasProduct: isGasProductMaterial(variationReaction, mh.material),
-  });
-
-  if (!field.unitToggle) {
-    return input;
-  }
-
-  /*
-  NumeralInputWithUnitsCompo copies `metricPrefix` into local state when it mounts and never reads
-  that prop again, so the unit picked in the column header would otherwise reach the material but
-  not the input showing it. Keying on the column's unit remounts the input, which then seeds itself
-  from the material that the header has just updated.
-  */
-  return (
-    <React.Fragment key={columnUnits[colId] ?? ''}>
-      {input}
-    </React.Fragment>
-  );
-};
-
-MaterialFieldCell.propTypes = {
-  data: PropTypes.shape({
-    idx: PropTypes.number.isRequired,
-    data: PropTypes.instanceOf(Reaction).isRequired,
-  }).isRequired,
-  colId: PropTypes.string.isRequired,
-  matGroup: PropTypes.string.isRequired,
-  sampleIdx: PropTypes.number.isRequired,
-  field: PropTypes.shape({
-    render: PropTypes.func.isRequired,
-    unitToggle: PropTypes.object,
-  }).isRequired,
-};
-
-/*
-Renders one reaction-level input of one variation. Unlike the material cells this needs no
-MaterialHandler - the scheme tab's reaction fields all work off the reaction plus the row's
-ReactionUpdateHandler.
-*/
-const ReactionFieldCell = ({ data, field }) => {
-  const { getRowHandler } = useContext(VariationsGridContext);
-  const reaction = data?.data ?? null;
-
-  if (!reaction) {
-    return null;
-  }
-  if (field.requiresNonInteraction && reaction.isInteractionReaction()) {
-    return null;
-  }
-
-  return field.render(reaction, getRowHandler(data));
-};
-
-ReactionFieldCell.propTypes = {
-  data: PropTypes.shape({
-    idx: PropTypes.number.isRequired,
-    data: PropTypes.instanceOf(Reaction).isRequired,
-  }).isRequired,
-  field: PropTypes.shape({
-    render: PropTypes.func.isRequired,
-    requiresNonInteraction: PropTypes.bool,
-  }).isRequired,
-};
 
 const AnalysesLinkCell = ({ data }) => {
   const {
@@ -581,15 +169,6 @@ GroupToggleButton.propTypes = {
 };
 
 /*
-Parent-header column picker. Every group carries one, so any single column can be hidden from the
-header it sits under. Hiding goes through the grid API rather than through the column definitions,
-so a toggle does not rebuild the columns and throw away the user's resizing; `hiddenColumns` is kept
-as state only to drive the checkboxes and to seed `hide` when the columns really are rebuilt.
-
-The popover is portalled to the body because the grid header clips its own overflow, which would cut
-an inline menu off.
-*/
-/*
 Header of a movable column. AG Grid drags the whole header cell, so the handle is an affordance
 rather than the drag source; it is revealed on hover by the stylesheet.
 */
@@ -605,129 +184,14 @@ DraggableHeader.propTypes = {
 };
 
 /*
-The cycles the gas phase unit buttons walk through, mirroring the conversions of convertTime and
-convertTemperature. Kelvin also occurs as '°K', which is not in the cycle and so falls back to its
-first entry - the very step the conversion makes of it as well.
+Parent-header column picker. Every group carries one, so any single column can be hidden from the
+header it sits under. Hiding goes through the grid API rather than through the column definitions,
+so a toggle does not rebuild the columns and throw away the user's resizing; `hiddenColumns` is kept
+as state only to drive the checkboxes and to seed `hide` when the columns really are rebuilt.
+
+The popover is portalled to the body because the grid header clips its own overflow, which would cut
+an inline menu off.
 */
-const GAS_UNIT_CYCLES = {
-  time: ['h', 'm', 's'],
-  temperature: ['°C', '°F', 'K'],
-};
-
-/*
-What a unit switchable column is currently in: the metric prefix for the amount fields, whose unit
-is fixed, and the unit itself for the gas phase fields, which have no prefix.
-*/
-const materialUnitOf = (material, unitToggle) => (
-  unitToggle.gasField
-    ? material.gas_phase_data?.[unitToggle.gasField]?.unit ?? ''
-    : unitToggle.prefixOf(material, isSbmmSample(material))
-);
-
-const unitToggleLabel = (unit, unitToggle) => (
-  unitToggle.gasField ? unit : `${metPrefSymbols[unit] ?? ''}${unitToggle.unit}`
-);
-
-const nextUnitOf = (unit, unitToggle) => {
-  const cycle = unitToggle.gasField ? GAS_UNIT_CYCLES[unitToggle.gasField] : unitToggle.prefixes;
-  return cycle[(cycle.indexOf(unit) + 1) % cycle.length];
-};
-
-/*
-Header of a column whose cells carry a unit switch, e.g. mass or volume.
-
-In the scheme tab a unit is switched one material at a time, which in a grid of variations means
-clicking the very same button once per row. The header button does that in one go: it advances every
-row of the column by one step of the same cycle the cell buttons walk through, so a column that is
-in one unit throughout stays that way.
-
-The unit is held in the grid instead of being read back from the materials, because the inputs copy
-their prefix into local state when they mount and never look at that prop again - see `columnUnits`
-in VariationSchemaTable for how the cells are made to pick a new one up.
-*/
-const MaterialUnitHeader = ({
-  displayName, colId, matGroup, sampleIdx, unitToggle
-}) => {
-  const {
-    variations, getRowHandler, columnUnits, setColumnUnit
-  } = useContext(VariationsGridContext);
-
-  const materialOf = (variation) => variation.data?.[matGroup]?.[sampleIdx] ?? null;
-  const rows = variations.filter((variation) => {
-    const material = materialOf(variation);
-    if (!material) {
-      return false;
-    }
-    // The gas phase cells render nothing unless the row's own product is the gaseous one.
-    return unitToggle.gasField ? isGasProductMaterial(variation.data, material) : true;
-  });
-
-  const firstMaterial = rows.length ? materialOf(rows[0]) : null;
-  // Until the header has been used, the column shows whatever unit its first row brought along.
-  const unit = columnUnits[colId] ?? (firstMaterial ? materialUnitOf(firstMaterial, unitToggle) : null);
-
-  const switchUnits = () => {
-    const nextCoUnit = nextUnitOf(unit, unitToggle);
-    setColumnUnit(colId, nextCoUnit);
-
-    rows.forEach((variation) => {
-      const material = materialOf(variation);
-      const mh = new MaterialHandler({
-        index: sampleIdx,
-        material,
-        reaction: variation.data,
-        materialGroup: matGroup,
-        onChange: getRowHandler(variation).handleMaterialsChange,
-      });
-
-      if (unitToggle.gasField) {
-        // The reducer converts the value and derives the next unit from the one handed to it, so
-        // every row advances from whatever unit it happens to be in.
-        const { value, unit: rowUnit } = mh.getFieldData(unitToggle.gasField, material.gas_phase_data || {}, unit);
-        mh.handler.gasFieldsUnitsChanged({ metricUnit: rowUnit, value }, unitToggle.gasField);
-        return;
-      }
-
-      mh.handler.metricsChange({
-        metricUnit: unitToggle.unit,
-        metricPrefix: nextCoUnit,
-      });
-    });
-  };
-
-  return (
-    <div className="d-flex align-items-center gap-1 w-100">
-      <DragHandle />
-      <span className="text-truncate">{displayName}</span>
-      {unit && (
-        <Button
-          variant="light"
-          size="xsm"
-          className="variations-unit-switch ms-auto py-0 px-1 flex-shrink-0"
-          title={`Switch the unit of "${displayName}" in every variation`}
-          disabled={!rows.some((variation) => permitOn(variation.data))}
-          onClick={switchUnits}
-        >
-          {unitToggleLabel(unit, unitToggle)}
-        </Button>
-      )}
-    </div>
-  );
-};
-
-MaterialUnitHeader.propTypes = {
-  displayName: PropTypes.string.isRequired,
-  colId: PropTypes.string.isRequired,
-  matGroup: PropTypes.string.isRequired,
-  sampleIdx: PropTypes.number.isRequired,
-  unitToggle: PropTypes.shape({
-    unit: PropTypes.string,
-    prefixes: PropTypes.arrayOf(PropTypes.string),
-    prefixOf: PropTypes.func,
-    gasField: PropTypes.string,
-  }).isRequired,
-};
-
 const ColumnVisibilityHeader = ({ displayName, columns, movable }) => {
   const { hiddenColumns, setColumnsHidden } = useContext(VariationsGridContext);
   const colIds = columns.map((column) => column.colId);
@@ -792,12 +256,10 @@ ColumnVisibilityHeader.propTypes = {
   movable: PropTypes.bool.isRequired,
 };
 
-const buildColumnGroups = ({
-  maxNumberOfSamples, showLoadingColumn, showGasColumns
-}) => {
+const buildColumnGroups = (variations, currentSegment, segmentFields) => {
   // The row-level columns are a group of their own so that they get the same per-column picker in
   // their parent header as everything else.
-  const groups = [
+  const group = [
     {
       groupId: VARIATION_GROUP,
       headerName: 'Variation',
@@ -828,9 +290,9 @@ const buildColumnGroups = ({
       ],
     },
     /*
-    Analyses are a group of their own rather than part of Variation: they get their own header, and
-    with it their own column picker and drag handle, so they can be hidden or moved independently.
-    */
+     Analyses are a group of their own rather than part of Variation: they get their own header, and
+     with it their own column picker and drag handle, so they can be hidden or moved independently.
+     */
     {
       groupId: ANALYSES_GROUP,
       headerName: 'Analyses',
@@ -845,58 +307,13 @@ const buildColumnGroups = ({
     },
   ];
 
-  MAT_GROUPS.forEach((matGroup) => {
-    const numberOfSamples = maxNumberOfSamples[matGroup] || 0;
-    const fields = FIELDS_BY_GROUP[matGroup];
-
-    for (let sampleIdx = 0; sampleIdx < numberOfSamples; sampleIdx += 1) {
-      groups.push({
-        groupId: `${matGroup}${GROUP_ID_SEPARATOR}${sampleIdx}`,
-        headerName: `${MAT_GROUP_TITLES[matGroup]} ${sampleIdx + 1}`,
-        matGroup,
-        columns: fields(showGasColumns.some((x) => x[sampleIdx]))
-          .filter((field) => !field.requiresLoadingColumn || showLoadingColumn)
-          .map((field) => {
-            const colId = `${matGroup}_${sampleIdx}_${field.key}`;
-
-            return {
-              colId,
-              headerName: field.header,
-              width: field.width,
-              ...(field.sticky ? { cellClass: STICKY_NAME_CLASS } : {}),
-              // Overrides the plain draggable header of buildColumnDefs with one that also carries
-              // the column wide unit switch.
-              ...(field.unitToggle ? {
-                headerComponent: MaterialUnitHeader,
-                headerComponentParams: {
-                  colId, matGroup, sampleIdx, unitToggle: field.unitToggle
-                },
-              } : {}),
-              cellRenderer: MaterialFieldCell,
-              cellRendererParams: {
-                colId, matGroup, sampleIdx, field
-              },
-            };
-          }),
-      });
-    }
-  });
-
-  // The reaction-level inputs sit after the materials, following the scheme tab's reading order.
-  groups.push({
-    groupId: REACTION_FIELDS_GROUP,
-    headerName: 'Reaction',
-    columns: REACTION_FIELDS.map((field) => ({
-      colId: `reaction_${field.key}`,
-      headerName: field.header,
-      headerTooltip: field.header,
-      width: field.width,
-      cellRenderer: ReactionFieldCell,
-      cellRendererParams: { field },
-    })),
-  });
-
-  return groups;
+  // "Schema" shows the reaction scheme; every other entry of the picker is a segment klass, whose
+  // editable fields become the columns instead.
+  return group.concat(
+    currentSegment === 'Schema'
+      ? schemaBuildColumnGroups(variations)
+      : segmentBuildColumnGroups(currentSegment, segmentFields)
+  );
 };
 
 const buildColumnDefs = (columnGroups, hiddenColumns) => columnGroups.map((group) => {
@@ -984,7 +401,8 @@ const VariationSchemaTable = ({
   allReactionAnalyses,
   reactionShortLabel,
   reactionId,
-  editMode
+  editMode,
+  reactionSegments,
 }) => {
   /*
   Seeded from the stored layout so the very first column definitions already carry the right `hide`
@@ -1007,9 +425,29 @@ const VariationSchemaTable = ({
   }, []);
   const [activeSlot, setActiveSlot] = useState(null);
   const [groupOrder, setGroupOrder] = useState([]);
+  const [currentSegment, setCurrentSegment] = useState('Schema');
+  const [allSegment, setAllSegment] = useState([]);
   const gridApiRef = useRef(null);
   const gridElementRef = useRef(null);
   const restoredRef = useRef(false);
+
+  useEffect(() => {
+    let cancelled = false;
+
+    async function load() {
+      const result = await getReactionSegments(reactionSegments);
+
+      if (!cancelled) {
+        setAllSegment({ Schema: {}, ...result });
+      }
+    }
+
+    load();
+
+    return () => {
+      cancelled = true;
+    };
+  }, [reactionSegments]);
 
   /*
   Mirrors the grid's own top level header order into state, so the toolbar always shows the groups
@@ -1141,28 +579,7 @@ const VariationSchemaTable = ({
     });
   }, []);
 
-  /*
-  Recomputed on every render on purpose: the parent mutates the `variations` array in place and
-  re-sets the same reference, so any dependency array keyed on `variations` would go stale. The
-  loops are over a handful of variations, and `columnDefs` below is memoized on the resulting shape
-  so AG Grid still only rebuilds its columns when the shape actually changes.
-  */
-  const maxNumberOfSamples = Object.fromEntries(
-    MAT_GROUPS.map((matGroup) => [
-      matGroup,
-      variations.reduce((max, variation) => Math.max(max, variation.data?.[matGroup]?.length ?? 0), 0),
-    ])
-  );
-  const showLoadingColumn = variations.some((variation) => !!variation.data?.hasPolymers());
-  const showGasColumns = variations.map((variation) => (
-    (variation.data?.products ?? []).map(
-      (product) => isGasProductMaterial(variation.data, product)
-    )
-  ));
-
-  const columnGroups = buildColumnGroups({
-    maxNumberOfSamples, showLoadingColumn, showGasColumns
-  });
+  const columnGroups = buildColumnGroups(variations, currentSegment, allSegment[currentSegment]);
 
   /*
   The pinned material column only ever shows the material of the slot currently at the left edge, so
@@ -1245,28 +662,58 @@ const VariationSchemaTable = ({
   const fixedGroupsAdvanced = fixedGroups.map((group) => ({ group, ...getGroupAllHidden(group) }));
 
   return (
+
     <VariationsGridContext.Provider value={gridContext}>
-      {}
-      { editMode &&
+      {/*
+      The segment picker sits at the right end of the same row as the group buttons, and stays there
+      when the group buttons are gone: it selects what the grid shows rather than editing it, so it
+      is not part of the edit mode toolbar.
+      */}
       <div className="reaction-variations-grid__toolbar d-flex align-items-center flex-wrap gap-2 mb-2">
-        {fixedGroupsAdvanced.map(({ group, allHidden, colIds }) => (
-          <GroupToggleButton key={group.groupId} group={group} allHidden={allHidden} colIds={colIds} />
-        ))}
+        {editMode && (
+          <>
+            {fixedGroupsAdvanced.map(({ group, allHidden, colIds }) => (
+              <GroupToggleButton key={group.groupId} group={group} allHidden={allHidden} colIds={colIds} />
+            ))}
 
-        <ReorderableList
-          horizontal
-          items={movableGroupsAdvanced.filter(({ allHidden }) => !allHidden)}
-          getItemId={(group) => group.groupId}
-          onReorder={(reordered) => applyGroupOrder([...fixedGroups, ...reordered.map(({ group }) => group)])}
-          renderItem={({ group, allHidden, colIds }) =>
-            <GroupToggleButton group={group}  allHidden={allHidden} colIds={colIds} />}
+            <ReorderableList
+              horizontal
+              items={movableGroupsAdvanced.filter(({ allHidden }) => !allHidden)}
+              getItemId={(group) => group.groupId}
+              onReorder={(reordered) => applyGroupOrder([...fixedGroups, ...reordered.map(({ group }) => group)])}
+              renderItem={({ group, allHidden, colIds }) =>
+                <GroupToggleButton group={group}  allHidden={allHidden} colIds={colIds} />}
+            />
+
+            {movableGroupsAdvanced.filter(({ allHidden }) => allHidden).map(({ group, allHidden, colIds }) => (
+              <GroupToggleButton key={group.groupId} group={group} allHidden={allHidden} colIds={colIds} />
+            ))}
+          </>
+        )}
+        <div style={{
+          position: 'absolute',
+          right: 0,
+          top: 0,
+        }}>
+        <Select
+          className="ms-auto"
+          // Matches the small buttons it shares the row with; without a minimum the control would
+          // collapse onto its own text.
+          size="sm"
+          minWidth="180px"
+          options={Object.entries(allSegment).map(([label, value]) => ({ label, value }))}
+          value={
+            currentSegment && allSegment[currentSegment]
+              ? { value: allSegment[currentSegment], label: currentSegment }
+              : null
+          }
+          onChange={({ label }) => {
+            setCurrentSegment(label);
+          }}
+          isSearchable
         />
-
-        {movableGroupsAdvanced.filter(({ allHidden }) => allHidden).map(({ group, allHidden, colIds }) => (
-          <GroupToggleButton key={group.groupId} group={group} allHidden={allHidden} colIds={colIds} />
-        ))}
+        </div>
       </div>
-      }
       <div className="ag-theme-alpine reaction-variations-grid" ref={gridElementRef}>
         <AgGridReact
           columnDefs={columnDefs}
@@ -1355,10 +802,15 @@ VariationSchemaTable.propTypes = {
   })).isRequired,
   reactionShortLabel: PropTypes.string,
   reactionId: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+  reactionSegments: PropTypes.array.isRequired
 };
 
 VariationSchemaTable.defaultProps = {
   reactionShortLabel: '',
+};
+
+export {
+  STICKY_NAME_CLASS
 };
 
 export default VariationSchemaTable;
